+++ src/system.rs
@@ -813,6 +813,36 @@ macro_rules! system {
             }
         }
 
+        impl<D, U, V> $crate::lib::cmp::Ord for Quantity<D, U, V>
+        where
+            D: Dimension + ?Sized,
+            U: Units<V> + ?Sized,
+            V: $crate::num::Num + $crate::Conversion<V> + $crate::lib::cmp::Ord,
+        {
+            #[inline(always)]
+            fn cmp(&self, other: &Self) -> $crate::lib::cmp::Ordering {
+                self.value.cmp(&other.value)
+            }
+
+            #[inline(always)]
+            fn max(self, other: Self) -> Self {
+                Quantity {
+                    dimension: $crate::lib::marker::PhantomData,
+                    units: $crate::lib::marker::PhantomData,
+                    value: self.value.max(other.value),
+                }
+            }
+
+            #[inline(always)]
+            fn min(self, other: Self) -> Self {
+                Quantity {
+                    dimension: $crate::lib::marker::PhantomData,
+                    units: $crate::lib::marker::PhantomData,
+                    value: self.value.min(other.value),
+                }
+            }
+        }
+
         impl<D, Ul, Ur, V> $crate::lib::cmp::PartialEq<Quantity<D, Ur, V>> for Quantity<D, Ul, V>
         where
             D: Dimension + ?Sized,