+++ src/lib.rs
@@ -18,7 +18,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! uom = "0.16.0"
+//! uom = "0.17.0"
 //! ```
 //!
 //! and this to your crate root:
@@ -61,7 +61,7 @@
 //! ```toml
 //! [dependencies]
 //! uom = {
-//!     version = "0.16.0",
+//!     version = "0.17.0",
 //!     default-features = false,
 //!     features = [
 //!         "usize", "u8", "u16", "u32", "u64", # Unsigned integer storage types.