+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Added
  * [#17](https://github.com/iliekturtles/uom/issues/17) Implement `PartialEq`.
+ * [#50](https://github.com/iliekturtles/uom/issues/50) Implement `Eq`.
  * [#26](https://github.com/iliekturtles/uom/issues/26) Implement `num::Zero`.
  * [#35](https://github.com/iliekturtles/uom/issues/35) Implement `num::Saturating`.
  * [#37](https://github.com/iliekturtles/uom/issues/35) Implement `serde::Serialize` and